+++ example/main.dart
@@ -1,5 +1,5 @@
 import 'dart:convert';
-import 'person.dart';
+import 'models.dart';
 
 main() {
   var p1 = new Person()