+++ example/models.dart
@@ -2,10 +2,15 @@ library example.person;
 
 import 'package:serializable/serializable.dart';
 
-part 'person.g.dart';
+part 'models.g.dart';
 
 @serializable
 class Person extends _$PersonSerializable {
   int id;
   String name;
+}
+
+@serializable
+class ClassWithMethod extends _$ClassWithMethodSerializable {
+  void someMethod(String p1) {}
 }