+++ test/serializable_test.dart
@@ -12,8 +12,8 @@ import 'package:test/test.dart';
 
 void main() {
   test('generated files', () async {
-    expect(await generate(await new File('example/person.dart').readAsString()),
-        await new File('example/person.g.dart').readAsString());
+    expect(await generate(await new File('example/models.dart').readAsString()),
+        await new File('example/models.g.dart').readAsString());
   });
 }